Tools need to use the Document that corresponds to the canvas passed into their mouse handlers, rather than relying on ActiveDocument.
The issue is when you have 2 canvases side by side in separate tabs, and one is selected.  Then you start drawing (MouseDown) in the unselected one.

The canvas gets to handle the mouse down first, before the tab control, so the canvas.DoMouseDown happens before the tab control makes the newly selected canvas the ActiveDocument.  Anything done in this initial MouseDown that uses ActiveDocument will be using the wrong document.

MouseMove and MouseUp don't suffer from this, but it's still better to use the canvas given to use rather than ActiveDocument.